Week 14 Work
I got confused and mixed up the weeks. The first submission with no note is in fact week 12, the second one with the incorrect note is week 13, and this is week 14. All the code should work but I was having internet issues while doing week 14 so I wasn't able to actually see if the code was working at the end.