Final project
Finding descriptions and really going through the references on the p5 website helped a bunch. For just about every line of code I checked the references to see what different ways I could do things.

There was a lot of trial and error with this, as some keywords worked with others, but not all others. I have this issue where my code works on my end, but not on yours, so I'm turning this in early-ish to hopefully get some feedback in that area. Thanks for a good semester!